manage multi-line texts
The texts can embed carriage returns, coded as \n strings into a one
line string. The behavior of the plotter is to split the text and issue
a series of paint event for each line. It is not the task of the painter
to interpret this and the painter can expect that text paint commands
only refer to single line of text at a time.