Change coord and relcoord type declaration
Following a discussion on LinuxFr, the declaration of a type in form

```
type foo = Foo of int*int
```

is different than

```
type foo = Foo of (int*int)
```

The former declaration is more compact in memory, because the type is
essentially equal to `int*int`, whereas the second declaration is a
pointer type on a pair type. The downside is that the pair type is no
longer autonomous and we can not construct or pattern match with a pair.