Fix illegal characters in cache.lib
Starting with version 5.0, the components are tracked with their
original library file in sch files. The components origin is now in
the format `library:component`. This means that the names in libraries
can not contain colons. But the cache library is a transposition of
the ones appearing in the sch files, so the colons are converted to
underscores.