Fix bug with keep option.
A bug was introduced with recursive search where the tmp files would
be moved to the root of filesystem. But anyway, using rename does not
work across mount points so this option is still brittle.

We open the diff files with different filenames depending on whether
the file needs to be kept or not.

Fixes #39